Date 03/26/2025

+++ src/main/java/com/ecommerce/belgro/Model/OrderItem.java
@@ -27,9 +27,9 @@ public class OrderItem {
 
     private int quantity;
 
-    private Integer mrpPrice;
+    private double mrpPrice;
 
-    private Integer sellingPrice;
+    private double sellingPrice;
 
     private Long userId;
